fix: canary scss and settings imports
