+++ packages/core/package.json
@@ -25,6 +25,8 @@
   },
   "devDependencies": {
     "@carbon/grid": "10.19.0",
+    "@carbon/ibm-cloud-cognitive-experimental": "^0.20.12",
+    "@carbon/ibm-cloud-cognitive-security": "^0.4.5",
     "@carbon/icons-react": "10.24.0",
     "@carbon/import-once": "10.5.0",
     "@carbon/layout": "10.17.0",
