+++ packages/experimental/src/components/_index.scss
@@ -10,6 +10,7 @@
 @import './AboutModal/index';
 @import './ActionBar/index';
 @import './BreadcrumbWithOverflow/index';
+@import './_Canary/canary';
 @import './EmptyState/index';
 @import './ExampleComponent/index';
 @import './ImportModal/index';
